# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API provides go template-microservice service",
        "title": "go-template-microservice API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/messages/sent": {
            "get": {
                "description": "Retrieves a list of sent messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "List Sent Messages",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Maximum number of messages to retrieve",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_internal_resources_response.SentMessagesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_pkg_utils.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/start": {
            "post": {
                "description": "Starts the message sending scheduler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Start Message Scheduler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_pkg_utils.HTTPSuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_pkg_utils.HTTPErrorResponse"
                        }
                    }
                }
            }
        },
        "/messages/stop": {
            "post": {
                "description": "Stops the message sending scheduler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messages"
                ],
                "summary": "Stop Message Scheduler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_pkg_utils.HTTPSuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-template-microservice_pkg_utils.HTTPErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-template-microservice_internal_resources_response.SentMessageResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "external_message_id": {
                    "type": "string"
                },
                "message_id": {
                    "type": "integer"
                },
                "sent_at": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "go-template-microservice_internal_resources_response.SentMessagesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-template-microservice_internal_resources_response.SentMessageResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "go-template-microservice_pkg_utils.ErrorSchema": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "go-template-microservice_pkg_utils.HTTPErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/go-template-microservice_pkg_utils.ErrorSchema"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "go-template-microservice_pkg_utils.HTTPSuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        }
    }
}